Added health bars for player units #11

+++ core/src/hr/fer/zemris/zavrsni/rts/objects/units/player/PlayerUnit.java
@@ -10,7 +10,7 @@ public abstract class PlayerUnit extends Unit {
     private boolean isSelected = false;
 
     public PlayerUnit(Animation<TextureRegion> animation, ILevel level, float width, float height,
-                      float defaultSpeed, float maxHealth, float attackRange, float attackPower) {
+                      float defaultSpeed, int maxHealth, float attackRange, int attackPower) {
         super(animation, level, width, height, defaultSpeed, maxHealth, attackRange, attackPower);
     }
